fix: categories filter must leave other filters intact
When the categories spliced the filter array, the "deleteCount" argument was missing, so it deleted
the remainder of the array, instead of just the matching "categories" filter, which is added back
with a different term.

This fixes so that it only deletes 1 element from the array and other filters are left intact
